+++ src/main/java/br/com/zalf/prolog/webservice/autenticacao/AutenticacaoService.java
@@ -3,11 +3,12 @@ package br.com.zalf.prolog.webservice.autenticacao;
 import br.com.zalf.prolog.webservice.Injection;
 import br.com.zalf.prolog.webservice.commons.util.Log;
 import br.com.zalf.prolog.webservice.commons.util.ProLogDateParser;
-import br.com.zalf.prolog.webservice.interceptors.auth.authenticator.StatusSecured;
+import br.com.zalf.prolog.webservice.interceptors.auth.ColaboradorAutenticado;
 import org.jetbrains.annotations.NotNull;
 
 import java.sql.SQLException;
 import java.util.Arrays;
+import java.util.Optional;
 
 /**
  * Classe AutenticacaoService responsavel por comunicar-se com a interface DAO.
@@ -20,7 +21,7 @@ public class AutenticacaoService {
     Autenticacao insertOrUpdate(@NotNull final Long cpf) {
         try {
             return dao.insertOrUpdate(cpf);
-        } catch (SQLException e) {
+        } catch (final SQLException e) {
             Log.e(TAG, String.format("Erro ao inserir o token para o cpf: %d", cpf), e);
             return new Autenticacao(Autenticacao.ERROR, cpf, "-1");
         }
@@ -29,47 +30,54 @@ public class AutenticacaoService {
     public boolean delete(@NotNull final String token) {
         try {
             return dao.delete(token);
-        } catch (SQLException e) {
+        } catch (final SQLException e) {
             Log.e(TAG, String.format("Erro ao deletar o token: %s", token), e);
             return false;
         }
     }
 
-    public boolean verifyIfTokenExists(@NotNull final String token, final boolean apenasUsuariosAtivos) {
+    @NotNull
+    public Optional<ColaboradorAutenticado> verifyIfTokenExists(@NotNull final String token,
+                                                                final boolean apenasUsuariosAtivos) {
         try {
             return dao.verifyIfTokenExists(token, apenasUsuariosAtivos);
-        } catch (SQLException e) {
-            Log.e(TAG, String.format("Erro ao verificar se o token existe: %s", token), e);
-            return false;
+        } catch (final Throwable throwable) {
+            Log.e(TAG, String.format("Erro ao verificar se o token existe: %s", token), throwable);
+            // Optamos por não tratar com o handler do Prolog, pois não queremos mandar nenhuma mensagem
+            // aos usuários nesse caso.
+            throw new RuntimeException(throwable);
         }
     }
 
-    public boolean verifyIfUserExists(@NotNull final Long cpf,
-                                      @NotNull final String dataNascimento,
-                                      final boolean apenasUsuariosAtivos) {
+    @NotNull
+    public Optional<ColaboradorAutenticado> verifyIfUserExists(@NotNull final Long cpf,
+                                                               @NotNull final String dataNascimento,
+                                                               final boolean apenasUsuariosAtivos) {
         try {
             return dao.verifyIfUserExists(
                     cpf,
                     ProLogDateParser.toLocalDate(dataNascimento),
                     apenasUsuariosAtivos);
-        } catch (SQLException e) {
+        } catch (final Throwable throwable) {
             Log.e(TAG, String.format("Erro ao verificar se o usuário com os seguintes dados existe: cpf - %s |" +
-                    " Data de Nascimento - %s", cpf, dataNascimento), e);
-            return false;
+                    " Data de Nascimento - %s", cpf, dataNascimento), throwable);
+            // Optamos por não tratar com o handler do Prolog, pois não queremos mandar nenhuma mensagem
+            // aos usuários nesse caso.
+            throw new RuntimeException(throwable);
         }
     }
 
     @NotNull
-    public StatusSecured userHasPermission(@NotNull final String token,
-                                           @NotNull final int[] permissions,
-                                           final boolean needsToHaveAllPermissions,
-                                           final boolean apenasUsuariosAtivos) {
+    public Optional<ColaboradorAutenticado> userHasPermission(@NotNull final String token,
+                                                              @NotNull final int[] permissions,
+                                                              final boolean needsToHaveAllPermissions,
+                                                              final boolean apenasUsuariosAtivos) {
         try {
             return dao.userHasPermission(token, permissions, needsToHaveAllPermissions, apenasUsuariosAtivos);
         } catch (final Throwable t) {
             Log.e(TAG,
                     String.format("Erro ao verificar se o usuário com o token: %s tem acesso as permissões: %s |" +
-                            " needsToHaveAllPermissions/apenasUsuariosAtivos: %b/%b",
+                                    " needsToHaveAllPermissions/apenasUsuariosAtivos: %b/%b",
                             token,
                             Arrays.toString(permissions),
                             needsToHaveAllPermissions,
@@ -81,11 +89,11 @@ public class AutenticacaoService {
     }
 
     @NotNull
-    public StatusSecured userHasPermission(final long cpf,
-                                           @NotNull final String dataNascimento,
-                                           @NotNull final int[] permissions,
-                                           final boolean needsToHaveAllPermissions,
-                                           final boolean apenasUsuariosAtivos) {
+    public Optional<ColaboradorAutenticado> userHasPermission(final long cpf,
+                                                              @NotNull final String dataNascimento,
+                                                              @NotNull final int[] permissions,
+                                                              final boolean needsToHaveAllPermissions,
+                                                              final boolean apenasUsuariosAtivos) {
         try {
             return dao.userHasPermission(
                     cpf,
@@ -93,7 +101,7 @@ public class AutenticacaoService {
                     permissions,
                     needsToHaveAllPermissions,
                     apenasUsuariosAtivos);
-        } catch (Throwable t) {
+        } catch (final Throwable t) {
             Log.e(TAG, String.format("Erro ao verificar se o usuário com o cpf/Nascimento: %d/%s tem acesso as " +
                             "permissões: %s | needsToHaveAllPermissions/apenasUsuariosAtivos: %b/%b",
                     cpf,