+++ src/main/java/br/com/zalf/prolog/webservice/interceptors/auth/authenticator/BearerAuthenticator.java
@@ -2,10 +2,12 @@ package br.com.zalf.prolog.webservice.interceptors.auth.authenticator;
 
 import br.com.zalf.prolog.webservice.autenticacao.AutenticacaoService;
 import br.com.zalf.prolog.webservice.commons.util.Log;
+import br.com.zalf.prolog.webservice.interceptors.auth.ColaboradorAutenticado;
 import org.jetbrains.annotations.NotNull;
 
 import javax.ws.rs.ForbiddenException;
 import javax.ws.rs.NotAuthorizedException;
+import java.util.Optional;
 
 public final class BearerAuthenticator extends ProLogAuthenticator {
     private static final String TAG = BearerAuthenticator.class.getSimpleName();
@@ -14,28 +16,37 @@ public final class BearerAuthenticator extends ProLogAuthenticator {
         super(service);
     }
 
+    @NotNull
     @Override
-    public void validate(@NotNull final String value,
-                         @NotNull final int[] permissions,
-                         final boolean needsToHaveAllPermissions,
-                         final boolean considerOnlyActiveUsers) throws NotAuthorizedException, ForbiddenException {
+    public ColaboradorAutenticado validate(@NotNull final String value,
+                                           @NotNull final int[] permissions,
+                                           final boolean needsToHaveAllPermissions,
+                                           final boolean considerOnlyActiveUsers) {
         Log.d(TAG, "Token: " + value);
         if (permissions.length == 0) {
-            if (!service.verifyIfTokenExists(value, considerOnlyActiveUsers))
+            final Optional<ColaboradorAutenticado> colaboradorAutenticado =
+                    service.verifyIfTokenExists(value, considerOnlyActiveUsers);
+            if (!colaboradorAutenticado.isPresent()) {
                 throw new NotAuthorizedException("Autenticação inválida, usuário não encontrado");
+            }
+            return colaboradorAutenticado.get();
         } else {
-            final StatusSecured statusSecured = service.userHasPermission(
+            final Optional<ColaboradorAutenticado> optional = service.userHasPermission(
                     value,
                     permissions,
                     needsToHaveAllPermissions,
                     considerOnlyActiveUsers);
-
-            switch (statusSecured){
+            if (!optional.isPresent()) {
+                throw new NotAuthorizedException("Autenticação inválida, usuário não encontrado");
+            }
+            final ColaboradorAutenticado colaboradorAutenticado = optional.get();
+            switch (colaboradorAutenticado.getStatusSecured()) {
                 case TOKEN_INVALIDO:
                     throw new NotAuthorizedException("Autenticação inválida, usuário não encontrado");
                 case TOKEN_OK_SEM_PERMISSAO:
                     throw new ForbiddenException("Usuário não tem permissão para utilizar esse método");
             }
+            return colaboradorAutenticado;
         }
     }
 }